Use long int for 16-bit compatibility in opt49.c
[git-p4: depot-paths = "//depot/openwatcom/": change = 31847]

+++ bld/ctest/regress/positive/source/opt49.c
@@ -4,7 +4,7 @@
 
 #include "fail.h"
 
-int test_function( int a )
+int test_function( long int a )
 {
     return( (a & 0xffff0001) != 0xffff0001);
 }
